Remove blue highlight from Link in ContextItem
