+++ src/Global/organisms/MobileNavbar.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import styled  from 'styled-components';
-import { Link } from 'react-router-dom';
 import { Tabs, TabContent, MobileTab } from '../../Tabs/index';
 import { TabListWrapper, TabList } from '../../Tabs/TabList';
 import { IMenu } from '..';
-import SvgLogoMark from '../../svg/LogoMark';
+
 import CloseButton from '../atoms/CloseButton';
 import MobileNavbarContainer from '../atoms/MobileNavbarContainer';
 import MobileMenu from './MobileMenu';
+import MobileLogoLink  from '../atoms/MobileLogoLink';
 
 const Container = styled.div``;
 
@@ -22,23 +22,6 @@ const HeaderContainer = styled.div`
   }
 `;
 
-const Logo = styled(Link)`
-  display: flex;
-  flex-basis: 0;
-  flex-grow: 1;
-  justify-content: center;
-  align-items: center;
-  object {
-    pointer-events: none;
-    height: 25px;
-  }
-`;
-const LogoMark = styled.div`
-  margin-right: 15px;
-`;
-
-const SVGObject = styled.object``;
-
 interface OwnProps {
   closeText?: string
 }
@@ -47,19 +30,18 @@ type IMobileNavbar = OwnProps & IMenu;
 
 const MobileNavbar: React.FC<IMobileNavbar> = ({
   content,
-  home = "/",
+  home,
   logoMark,
   supportUrl,
   defaultMenuOpen = true,
   closeText,
 }) => {
+
   return (
     <Container>
       <Tabs>
         <HeaderContainer>
-          <Logo to={home}>
-            <LogoMark>{logoMark ? <SVGObject type='image/svg+xml' data={logoMark} /> : <SvgLogoMark />}</LogoMark>
-          </Logo>
+          <MobileLogoLink {...{home, logoMark}} closeId='closeMenu' />
           <TabList defaultTabId='closeMenu'>
             <MobileTab tabFor='notifications' icon='Notifications' />
             <MobileTab tabFor='user' icon='UserProfile' />
